Improve the performance of pbessolve_parallel
The simplify_rewriter has been adapted to reduce
to store intermediate terms.

A reverse iterator has been added for lists. This
reverse iterator stores the list element in an array
and then traverses them. This allows to construct
new lists in the same order as the old list. This has
been applied to split_finite_variables.h

This all has some beneficial influence on the performance of
pbessolve_parallel.