Repaired a few stupid errors in the simplifying rewriter.
The simplifying pbes rewriter was adapted to
evaluate the right hand side of the implication arrow before
the left hand side. This had as effect that a test
expression of the shape

    false => PHI

where PHI was a complex formula, would not rewrite to true.
Moreover, a case was forgotten, leading to expressions
not(true) not being simplified to false.

All tests now succeed again in decent time.