Adapted the iterators in indexed sets and marking in the enumerator.
The iterators in indexed sets went forward from the end() iterator,
and subsequently backward to a valid position. This works on some
platforms, but not on all, causing failures.

In the enumerator, a designated mark function for garbage collection
did not check for undefined terms. This has now been adapted.

Both caused problems, but only when running the toolset in parallel.
In the first case most likely because data in memory could change while
moving forward beyond a container, and then backwards again. The
second might have been due to the fact that garbage collection can
take place at basically any moment, which does not happen with one
process.