Replaced optimized_and, ..._or, ..._not and ..._imp functions.
Replaced the optimized_XXX functions from

         term optimzed_XXX(arg1, arg2)

by

         void optimized_XX(result, arg1, arg2)

as the latter does not implicitly put a term on the
stack that is constructed and deleted.

An observation is that the CLANG compiler seems to
inline such optimized functions, avoiding construction
and destruction of terms, whereas the GCC compiler
(on the Mastodont) does not. This makes the parallel
toolset on the Mastodont much slower than on Apple.

All tests succeed.