Added an explicit flag to compile the toolset in parallel or non parallel mode.
By uncommenting

in the file mcrl2/utilities/configuration.h the toolset is compiled
for parallel use. Otherwise, it is used sequentially. In the sequential
mode there are no thread_local variables and the busy/forbidden protocol
to safeguard the parallel construction of terms is switched off.

The system uses garbage collection by protection sets, though.