Adapted the example for Othello.
The othello/reversi example had become quite incomprehensible,
and has been restored to its older form. Moreover, a running
script has been added that applies lpsparunfold, to replace
the List of List representing the board by individual board
positions in the lps. This did work for the example as it
existed a long time ago, but not anymore for the current
example. Using this, a board of 4x4 can be solved, but
larger boards, even a 5x5 board is a challenge with current
tools. A 4x5 board is doable.

Othello or reversi is interesting because with each move the
board can change a lot (as enclosed pieces change color). This
makes this example pretty hard for symbolic tools. Actually,
symbolic tools, such as lpsreach, seem to completely choke
on even small examples.