Improved the performance of the compiling rewriter.
In some cases places were reserved on the rewrite stack, where this
was not necessary. At other places, unprotected_assign has been used
which copies a term to a new position. If the original position is
stable, i.e., will not be overwritten, before a next lock_shared
takes place, it does not require a lock_shared to move the data,
which is provided by the unprotected assign.