Correctly reduce transitions of a LTSs with a single state -- fixes bug #1666
A LTS with a single state generally does not need to be minimised; however,
the input format allows to specify the same transition more than once,
and in that case minimisation removes the redundant transitions.  This
``abuse'' (if you will) of minimisation is now also offered by the standard
bisimulation algorithm.
Also added two test cases to librarytest_mcrl2_lts_ltsconvert_test that
test the correct behaviour.

I further took the opportunity to change comments so they accurately reference
the TACAS 2020 publication that describes the algorithm.