Print warning about exceeding time bound O(m log n) -- related to bug #1666
If a LTS has only n=1 state, minimising it with an O(m log n) algorithm
should take no time, as in that case log n=0.  However, occasionally a
user minimises such a LTS to remove duplicated transitions.  Then the
tool should issue a warning that the time bound may be exceeded.

To avoid confusing warnings, weak bisimulation minimisation of a LTS
with a single state is also adapted; for such an LTS, it is not necessary
to call (branching) bisimulation minimisation, as duplicated transitions
will be removed by other parts of weak bisimulation minimisation,
namely scc_reduce().