Repaired bug #1656, term_balanced trees should be aterms.
Repaired this problem by making all term_balanced_trees proper
aterm_appl's. A singleton tree is now an application of the function
@single_node@ to its argument. If a tree of numbers or lists
is created, then a term_balanced tree is now always a proper aterm_appl.

Furthermore, repaired some warnings, provide by the clang compiler on
the mac.

Some symbolic header tests do not succeed, but this seems to be an
unrelated problem.