Adapted the othello example.
The description of the Othello game was adapted and it turned out that
it took a long time to prove it deadlock free using a modal
formula. The reason for this is that in the adapted form
the expression saying that a move is allowed, is very big
when rewriting is applied to it. The example has now been
changed a little to make it more concise, and premature
rewriting has been prevented by adding a dummy condition
to one of the rewrite rules.