Repaired a failing test under garbage collection stress.
If garbage collection is set to "always do a garbage collection
when creating a term" three tests fail, namely:

      281 - librarytest_mcrl2_data_enumerator_test (Failed)
      295 - librarytest_mcrl2_data_rewriting_test (Failed)
      496 - librarytest_mcrl2_lts_lps2lts_test (Failed)

without jittyc and extended tooltests. This commit repairs the
first failing test, which is due to a too restricted test setup
and which does not indicate a problem in the toolset.