Added an APPLE compiler flag for arm64/x86_64 compilation.
On the new M1 apple computers with an arm processor,
it is possible to run x86_64 code, or arm64 code. A flag is added
to compile the toolset for either platform.

Unfortunately, the Sylvan code is very dependent on x86 code, including
the use of explicit x86 assembly instructions (sic, sic, sic, sic)
and for some reason everything appears to depend on this code, so the
toolset does not yet compile for the ARM platform.

However, the M1 macs execute x86_64 code faster than an x86 macs, and
hence generating x86 code is usable for the moment.