Add PBESSMT tools
This includes two tools, pbes2cvc4 and pbes2yices, that are the work of Ruud Koolen.

The techniques used by the tools are described in the following paper:

  Koolen, R.P.J., Willemse, T.A.C., and Zantema, H.
  Using SMT for Solving Fragments of Parameterised Boolean Equation Systems.
  In: Automated Technology for Verification and Analysis. pp. 14–30 Springer, Cham (2015).
  https://doi.org/10.1007/978-3-319-24953-7_3.

I am adding these tools as experimental tools to make sure that the tools still compile.
Tim Beurskens is planning to experiment with these tools in his graduation project.