Resolve bug #1685.
Added an explicit check to the mcrl2 typechecker saying that
if in a variable assignment of the shape P(x1=t1,...,xn=tn) all
implicitly mentioned assignments of the shape y=y, if y is also a
parameter of process p, should have a valid term y at its right
hand side.

Added a test to check that the typechecker now fails on the
example given in this bug report.