Solved bug report #1687
Hidden actions were not properly recognized in ltsconvert. Multiactions
that were partly hidden compared with the default action label, and had
to be compared with a tau-action. This caused multiple tau-actions to be
listed in the list of actions of an lts, and this caused other tools to
go astray.

The code to rename actions has completely be rewritten.

This should solve #1687.

A test have been added.