Removed lps2lts_parallel and pbessolve_parallel.
Integrated the functionality of lps2lts_parallel and pbessolve_parallel tools
in lps2lts, and pbessolve/pbes2bool. The only noticable difference
is that these tools now have an option --threads that can be used to
indicate how many threads are used to generate a statespace or solve a
pbes.

The experimental tools lps2lts_parallel and pbessolve_parallel have been
removed. New and old tools already shared the majority of the code, so,
there should not be any noticable difference when using the tools
with one thread, which is the default.