Update AddMCRL2Binary.cmake
Removed a useless dependency on GLOB. In case get_filename_component fails however, GLOB_RECURSE is used to attempt a fix.

Before it could happen that SOURCE_DIR = source/, and a file is added named a.cpp, but it resides in folder source/subfolder1/a.cpp and it would not be found by get_filename_component, and subsequently GLOB would also miss it.

Other solutions:
- Adding separate source directories in CMakeLists input like with include directories
- Explicitly stating the relative path from the source directory in cmakelists.txt (not tested)