[mcrl2ide] Changed the behaviour of the property widget when changing the property
When a property is saved or deleted, any verification process that runs for this property is now aborted.
This solves a confusing situation where one can change the property while verification is ongoing (by having the edit window open before verifying the property), after which the result of the verification is shown on the widget of the previous version of the property.

Also made it so that the edit and delete property buttons no longer become disabled when verifying the property, as the change mentioned above makes this unnecessary.