[mcrl2ide] Changed the behaviour when a property is successfully parsed
When clicking "Save and Parse" in the add/edit property dialog, instead of showing a pop-up dialog when a property is parsed successfully, a bit of text next to the button mentions that it is parsed successfully. This saves the user a click or button press.
When parsing is not successful, a pop-up dialog is still shown.