Resolved bug #1695. Assertion failure in parallel exploration.
The indexed_set data structure, which is heavily used in state space
generation and solving pbesses, is now accessible by different threads.
The threads must however provide their thread identifier.

At some places this thread identifier was not passed on, and hence
it defaulted to the thread number 0. So, at some points multiple threads
where accessing the busy-forbidden flags of thread 0 simultaneously and
this was noted in asserts.

The problem is resolved by only using the number 0 for threads that
are not parallel, and by counting parallel threads from 1 upwards.
Additional asserts have been added to catch -- in debug mode --
when thread numbers are not used as intended.

All tests succeed, and these assertion do not fail.

The found issue was serious and could have caused malbehaviour by
incorrectly counting states/pbes variables or dropping them altogether.
It is interesting that such behaviour was not observed by me.