Improved the performance of --cached for lps2lts with multiple threads.
When using --cached in lps2lts there was one mutex that protected every
access to the cache. This mutex was locked and unlocked for every access to every
summand. This was a major performance bottleneck.

Now the mutexes have been assigned per summand, provided local_caching,
which is the default, is used.

For a locomobile example of the Maeslantbarrier, I just saw 400M states
being generated in one hour (without saving the transition system) on a
20 procesor M1 mac. I expect that more speedups are possible.