Removed index in boolean_variable and propositional_variable_instantiation
When a variable, function_symbol, process_identifier, boolean_variable
and a propositional_variable_instantiation were generated, a unique
index is added. For boolean_variables and
propositional_variable_instantiation this was very time consuming
expecially in a parallel setting, where assigning a number was protected
by a mutex. This mutex appears to be the main cause that pbessolve could
not really be paralellized as all threads were queueing upo creating a
new number for a propositional_variable_instantiation.

Somewhat astonishing this index was not used at all. So, I removed
the generation of the index, speeding up pbessolve substantially,
especially with multiple processors.

Process identifiers also get index numbers, but these
numbers also do not appear to be used either.

All tests succeed.