Made the atermpp::unordered_map proper.
The atermpp::unordered_map did put the values stored in it also in the
protection set. This was not supposed to happen, as the elements of the
unordered map are being marked by traversing the atermpp::unordered_map
explicitly. Actually, the values in the atermpp::unordered map would
now be marked twice, once because the were part of the unordered map
and once because they were stored in a protection set.

There is a small issue that elements of an unordered_map can now
both be a of its own type T and of the type aterm_reference<T> and
these types are convertible to each other. This means that for elements
from an unordered_map it must be stated explicitly what their
actual type is. It would be nice if we can resolve this, such that
aterm_reference<T> will never be encountered by ordinary programmers.