Improved the performance of parallel state space generation.
When generating a state space on many processors, where the
computational effort per state is low, it turns out that the
level of parallelism is low. This for instance happens with the
compiling rewriter and the --cached option.

The problem lies in mutex variables that surround certain
parts. Recently, replacing replacing the single mutex for
all summand caches by a single cache per summand gave a very
notable performane boost.

It turns out that after this improvement the mutex surrounding
the single todo buffer wbich all processes use to store their
new states and obtain states to be investigated is now the culprit.
Therefore, each process now has its own todo buffer. When it finds
that the central todo buffer is empty, while there are processes
that are idle as their own todo buffer is also empty, processes
put 1/4th of their own states in the central todo buffer.
This has a notable performance benefit of factors.

Another change is that the mutex protecting processing transitions,
i.e. mainly writing transitions to disk, has been pushed towards the
lts-builders. This has as advantage that if nothing is done with
transitions, i.e., if they are not stored on disk, no protection
by mutexes is performed anymore.

Still, now linear speedup is achieved with many processors. As it
stands the mutexes used for caching in summands turn out to be the essential
culprits again.

For reference, below the current improvement is indicated for the
command
time lps2lts -vrjittyc --threads=X --cached 1394-fin-vlarge.lps --timings
1394-fin-vlarge.mcrl2 is the 1394 protocol with 10 data elements.
It has 7207497 states and 15643295 transitions.
In the left column the times of the previous release toolset is
provided. In the right column the new times are listed. Note that
the new version is substantially more parallel, but with many
processors, the degree of parallelism still needs to be improved.

   X
   1        57s     54s
   2        34s     32s
   4        25s     18s
   8        25s     12s
  16        28s     12s
  20        30s     12s

All tests succeed.