Improved the performance of state space generation.
By using a dedicated find function that does not calculate a key
when caching in a summand the performance of state space
generation using the --cached option is increased with appr. 10%.
For this purpose a dedicated equality function and a dedicated
hash function had to be added to the unordered_maps which forms
the summand caches.

Also for parallel state space generation there is some improvement,
but it is slightly less.