Removed index from variables and process variables.
The index has been removed from variables and process variables.
The latter were not used at all, and the former was only used
in a mutable_indexed_substitution. The mutable indexed substitution
has been replaced by a atermpp::utilities::unordered_map.

This simplifies the code, in the sense that it is not necessary
to assign a number to each variable that is being generated, which
is an expensive operation. A lookup in a mutable_indexed_subsitution
did become a little bit slower, as now a hashtable lookup must be
done, instead of a direct lookup in a table.

By and large the effect on the performance is quite neglectable
compared with the last commit.

All tests succeed.