Added an example by Hopcroft, which is claimed to be "hard".
John Hopcroft provides an example in his seminal paper, an n log n
algorithm for minimizing states in a finite automaton, nl. B(n)
of which he claims that it shows that his complexity n log n is
optimal. The example is not clearly described, but it appears that
a more efficient minimization scheme for this example exists.

But in any case it is an interesting example.