Sligthly improved performance of reading aterms from disk in binary format.
The performance of reading terms in binary format is slightly increased
by using make_ functions instead of functions passing the arguments on
the stack. This saves calls to "register_variable" and
"unregister_variable" that are used to put the memory position where the
intermediate result is stored in a protection_set.