Apparently C++20 is more relaxed with matching of operator==.
The compiling rewriter failed to compile, due to a:

 error: use of overloaded operator '==' is ambiguous

error message. In this commit it is resolved by adding a constant.
An investigation on the internet points to more relaxed rules
in C++20. I am not sure whether this is the real culprit, but
it appears to be resolved now.