Adapted the format of a probabilistic state.
A probabilistic state now consists either of a vector of
state/probability pairs or a single state with implicit probability 1.
This means that when reading many states that are not probabilistic
the probability 1 does not need to be stored. This has numerous
advantages, among others that tools like ltsconvert and ltsinfo
are faster.

The BAF version of objects stored in files as aterms has been increased
to 33543, because the .lts format has changed, and in the lasts weeks
also the way variables, process names and Bes variables are stored have
changed, too.

All tests succeed.