[mcrl2_gui] Disabled Shift+Enter in codeeditor
Shift+Enter created a new line, but in the editor it looked like this new line still belonged to the same line number, while this was not the case in the file. This was mainly an issue when the original line was commented out: in the editor it looked like following lines created with Shift+Enter were also commented out, while in reality they were not.