Allowed ltsconvert to also work sometimes on probabilistic ltss.
Probabilistic LTSs could not be handled by ltsconvert. As a result
it was not possible to translate a probabilistic .aut file to a .fsm
file, or similarly an probabilistic .lts file to say a probabilistic
.aut or .fsm file.

This was reported in bug report #1697.

With this commit this should have been repaired.

Note that the the equivalences only allow to transform non probabilistic
files. It is also not possible to determinise a probabilistic file, as
it is not clear what that means.

All tests succeed.