Resolved two parallelism issues with lps2lts.
When searching for divergencies and nondeterministic states,
the toolset crashed when running in parallel. The reason was
that both activities were not thread safe.

Searching for divergencies is now done sequentially, protected
by a mutex. This may be parallelized at some later time.

Searching for nondeterminism was done by setting a global
variable when a new state was visited, and storing the outgoing
transitions when visited in a set. These global variables
have now been replaced by a vector of sets, one set per thread.