Use find_equalities in sumelm
Eliminate this code duplication, by replacing an ad-hoc algorithm with
the find_equalities function that is also used in the
one-point-rewriter. With the recent changes to find_equalities, this
makes sumelm slightly more powerful. Split some code out of the
one-point-rewriter to facilitate re-use.

Behaviour of sumelm is changed slightly: a summand
sum d:D . (d == d1 && f) -> a . P()
is now simplified to
(d1 == d1 && f) -> a . P().
The old behaviour was to produce (f) -> a . P().