Made the --help-all flag visible in the help text.
It turned out that quite a number of people were not aware
of the --help-all flag for the tools that would allow to inspect
hidden and experimental options. Although these hidden options
are not intended for common use, some may still find them
useful. Now at least there is a comprehensible way to find
out that such options exist.