Avoided passing through a mutex when generating a fresh variable.
When a fresh function symbol was generated, necessary to make a fresh
variable, the function register_prefix was invoked, which checked
whether the function symbol was already in use. In register_prefix,
a mutex is used, forming a bottleneck for highly parallel state
space generations. The change that is made, is to store a shared
pointer to the information, and in this way the call to register_prefix
can be circumvented.

All tests succeed.