Repaired a subtle error where exceptions could disturb the busy/forbidden protocol.
Code that had to take place in the shared section of the busy/forbidden
protocol hitherto had the form:

  lock_shared();
  protected_code;
  unlock_shared();

But if the protected_code would throw an exception, the unlock_shared()
would be skipped. One of such exceptions is thrown when the rewrite
stack is full. It is caught by the rewriter, that resizes the stack and
starts rewriting the term anew. However, as at least one unlock_shared
would be skipped, the lock_depth in lock_shared would have become
larger than 1, meaning that this process would never relinguish the
busy_flag. As a consequence, when a process wanted to enter the
exclusive section, this would not be possible, blocking most of the
other processes. The net result is that only one process, the one
with non-zero lock-depth would continue processing, one process would
be busy-looping in an attempt to set the forbidden flag, and all
other processes would be locked in mutexes.

The solution is that the unlock_shared is now also executed explicitly
when and exception is thrown, using

    try ... catch (...) { unlock_shared(); throw; }