The code cleanup of explorer.h did cause too much code to be removed.
In the code cleanup of three weeks ago of explorer.h the computation
of a term that would serve as the key for cached data elements in
summands was accidentally removed. This caused the cached state space
generator to fail. The problem was reported by Pham Duc Minh.
The missing line is reinserted....