Unfold some pattern matching function applications
Whenever an application f(a1,...,an), where f is defined by pattern
matching, occurs within the scope of Det or pi, unfold it once when
updating LPS parameter updates and LPS initialisation.

This still needs to be extended with a heuristic to unfold f(a1,...,an)
multiple times, so that the Det and pi disappear in most scenarios.