Repaired highway search.
Highway search was damaged when making it parallel.
It has now been replaced with a more explicit construction
that has two todo buffers for states, per process. One for
the new states, and one for the states currently under exploration.
Each thread has its own max-todo buffer, which means that the highway
breadt is typically --todo-max times the number of threads.