Resolved a problem with parallel state space generation on ARM.
When running the state space generation in parallel on a 20 processor
ARM M1 processor, the number of states being generated appeared to
differ from run to run. It turned out to be due to the indexed set
where apparently, the order of some statements is crucial.
By adding one:

std::atomic_thread_fence(std::memory_order_seq_cst);

the problem is resolved. It appears that an acquire or release
fence can also be used, which theoretically should not be the
case.