Made the tools txt2res, txt2pres, prespp, respp, presinfo, resinfo.
It turns out that the syntax for bes's (and consequently for res's)
is not used in txt2bes. Instead txt2bes reads a pbes, and checks
whether the variables have no parameters. This means that the
bes (and res) syntax, as well as the bes internal class format
do not appear to be used. It is unclear how to deal with this.