Repaired the possibility to group domains.
When selecting domains to be grouped
in diagraphica, the menu was not always selectable.
This was due to the fact that when reselecting
items in the Domain window, the internal administration
was not adapted. This has now been mended, meaning that
it is possible to group domains again for more pleasant
visualisation. Problem was noted by Ferry Timmers.