This invariant does not hold so it has been removed.
Consider the case where a term is being rewritten inside the
transformer of make_appl(f, args, transformer). Then we enter
the shared section to build up arguments with the transformer applied.
In the transformer we start a new rewrite and at some point this
exception can be thrown, but then the depth is still greater than zero
since the outer term is being constructed.