A version that seems to work for smaller examples.
Using this version the example where a coin simulates a dice
and where all examples that were in the paper submitted to concur
could all be calculated with correct answers.

Much needs to be done, to let it work in general, and to improve
efficiency. For instance generating 4500 equations for the game
of the goose went well, but solving it, did not lead to anything,
yet.