Parallel processing of stochastic processes contained issues.
When generating stochastic processes in parallel, the initiatialisation
was not done properly, using thread index 0, where 1 should have been
used, which could cause a shared flag to be set, but not reset,
leading to the problem that only one process could continue, while
a second process would attempt to obtain an exclusive lock, leading
to 1) the situation where in general only one process would be
able to make progress. But this process would ultimately also
get stuck, because the indexed_set would not be able to make
new indices.

It appears that not all issues with generating stochastic state
spaces have been solved.