Added an explicit memory fence.
When generating a stochastic state space with multiple threads,
state space generation could fail as one thread would grap a
shared_lock and never give it back.

Investigation seems to show that this is due to a piece of code
of the shape:

   do
   {
     std::size_t index=m_hashtable[position];

     .. code not changing position ..
   } while (true);

The optimiser would not read index in each loop, but rely on an old
value. However, this was just a busy loop, waiting for another process
to change m_hasthable[position]; Typically, in some runs no problem
would occur, while other runs were problematic. Adding debug code would
typically lead to the situation where the problems would be gone.

By adding an explicit memory fence, the code appears to run reliably,
from which I conclude that this is the solution to the problem.