Sped up the lineariser.
The lineariser has not really been adapted to the new parallel framework
as it turns out to be fast enough. But on an example generated from CIF
it turned out that a few places could easily be sped up substantially.

All tests succeed in debug mode on an ARM processor.