Repaired a linearisation bug with stochastic processes.
When using stochastic processes linearisation could go astray
due to a somewhat stupid bug in the lineariser. The example was found
by Jari Martens when modelling the Maeslantbarrier control system to
establish a reasonable failure rate.

A minimized example is given below:

proc Environment =
        sum realReading: Nat . (realReading<2) ->
            a . EnvironmentSensorProcess;

proc EnvironmentSensorProcess =
        dist f:Bool[if(f, 1/4, 3/4)].f->b.delta;

init Environment;

When linearising the distribution of EnvironmentSensorProcess must be
removed and replaced by a distribution after the action a in
Environment. When replacing EnvironmentSensorProcess with a process with
an initial distribution in Environment, the resulting process was
ignored, and the original remained in place. This would only happen
when a sum operator would be present. The fix is simple.

A test has been added.