Reverted commit 565da0a in ltsgraph.
Commit 565da0a, which only consisted of changing some
int's into size_t, had as completely unexpected side effect
that the borders of states were not drawn anymore.

The purpose of the original commit was to remove some warnings
for comparisons between int's and size_t's. Some of these warnings
have now reappeared.

It is completely unclear why this phenomenon of disappearing state
borders is happening, as the code changes seem to be harmless....