Made the border of probabilistic states blue.
In ltsgraph the probabilistic states "stood out" by drawing a half-sfere
over them. The effect was that these states were big black nodes in 2D
and half colored states in 3D mode. This was not a good way to let them
stand out.

The probabilistic states are now made visible by coloring its border
blue, instead of black. It was already the case that probabilistic
states were blue internally, and this has now become visible again.
In the future a possibly better way could be designed to let
probabilistic states stand out, but for the moment the current approach
seems adequate, and better than the use of "half-spheres".