Resolved a problem in ltsgraph with label sizes.
The sizes of transition labels appeared to change in 2D mode
quite randomly. This was due to resizing these depending on the
distance from the camera. Because this distance was taken
as an integer the closest object would always be drawn slightly
bigger.

Labels in 2D are now not resized anymore depending on their
distance to the camera.