Repaired an issue with the initial distribution in lps.
Based on an example by Jari Martens it was revealed that a
distribution in the initial process could lead to issues.
Concretely, in

proc P(v:Nat)= dist e:Bool[if(cond(v),...,...)].....

init P(1);

the value of 1 for the parameter v was not substituted for v in the
distribution. This has now been mended, and a test has been added.

Moreover, the linearisation tests have been strengthened by checking
whether the linearised processes are well typed.

All tests succeed.