Worked on a faster implementation to solve RESs
Essentially the new implementation to solve RESs uses an
internal data structure. It can be activated by the --algorithm=numerical_directed flag.
A feature that is still to be added is that only equations of which a
variable in the rhs has changed should be evaluated. This should change
the quadratic nature of the current algorithm into a linear one.
For large presses the current evaluation takes too long, ie. hours
for 1M pres equations.