Removed two errors in probabilistic lps's.
The first error was caused by not renaming the
bound variables in distributions before generating the state space.

The second error was caused by incorrectly storing a probabilistic
state from a string in .aut format into a probabilistic state.
A single state with probability 1 must be stored separately.
If there are more states with probabilities <1 they must be stored
in a vector.

A test have been added to check for these problems.