Added the rewriter to action_rename.
The rewriter was not applied within the procedure to rename
actions, used within lpsactionrename. This could lead to
a situation, as shown by Kevin Jilissen, that an exponentional
blowup of the linear process would occur, which would subsequently
be reduced when the rewriter would be applied.

This exponential blowup is now avoided by using the rewriter
within action_rename.