Adapted some error messages when solving pbesses.
When pbesses were solved, this could result in a rather cryptic message
namely

Unexpected term " + expression + " encountered in Rplus

The cause of this was generally that a term could not be rewritten to
true of false, or quantification was done over a non enumerable domain,
such as for instance bags.

The error messages have now been replaced by something less cryptic,
avoiding confusion.

The problem was brought to my attention by Morris Boers.