Resolved a serious bug in resolving_name_clashes in modal formulas.
The function resolve_name_clashes in a modal formula would not properly
deal with equal bound variables which would be nested more than two
times. A typical example is:

     forall x:Nat. forall x:Nat. forall x:Nat.....

The problem was due to the fact that the erase function on multisets
would forget all names x when invoked once, whereas it had to remove
only one. This problem was already there since the start of adding
the function resolve_name_clashes, and may be the explanation for
some issues with bound variables within modal formulas.