Repaired clustering of linear processes with distributions.
When the variables in a distribution would coincide with the
variables in sum operators, the variables in the distribution
could be renamed to variables in the sum, after which
substitution would go astray, leading to a non bisimilar
clustered linear process.

This should have resolved by the simple fix that is committed.