Adapted lpsxsim and lpssim to be able to simulate probabilistic processes.
Lpssim and lpsxsim can now simulate probabilistic processes. In
principle the user alternatingly selects a probability and then
a transition. If a probability distribution has one state with
probability 1, then this state is automatically chosen. This
makes the behaviour for non-probabilistic processes indistinguishable
from the behaviour of the previous variants of lpssim and lpsxsim.

Both simulators now use the explorer made by Wieger, instead of
the next-state-generator, made by Ruud Koolen.

The tau-priorisation mode has been removed from lpsxsim, as
it was not used and could lead to infinite uninteruptable behaviour.

It is possible in lpsxsim to indicate that automatically a state must
be chosen from the probability distribution. Also in play mode
such states are automatically chosen, where the probability of the
transitions is respected.

All tests succeed.