Changed '=' into double backquote in a Sphinx documentation file.
For some reason the =-symbol surrounding a string had a meaning in
typesetting. But with the current version of Sphinx that is not the
case anymore. All occurrences of an = surrounding a string have been
replace by a double backquote. It needs to be seen whether this leads
to the desired typesetting. Maybe a single backquote, or other Sphinx
typesetting directives are more appropriate.