Strengthened the lineariser regarding distributions.
If an expression is encountered in the lineariser of
the form

    c -> dist d:D[f(d)].p

then this expression was transformed in some cases into

    dist d:D[f(d)](c -> p)

This is unproblematic, except if the distribution f(d) is
only valid in the context of the condition c. A concrete example
is:

   (m==2) -> dist n:Nat[if(n<m,1/2,0)].a(n).P

where the validity of the distribution depends on m==2, and
it is not valid if the condition is pushed inside.
Therefore, when pushing a condition inside, the distribution
is made dependent on c by yielding:

    dist d:D[iif(c,f(d),if(d==d_default,1,0))](c -> p)

At some places this transformation was already carried out in this
way. It now also has been added to the routine "distribute_condition".
The problem was spotted by Tim Willemse while modelling a cotton loom.

A test has been added.