Replaced every aterm and unprotected_aterm with aterm_core and unprotected_aterm_core
It is the intention to remove the class aterm and replace it with the
current class aterm_appl. Classes like aterm_int must then derive from
aterm_appl, which in this branch is already the case.

To achieve this, first aterm is renamed to aterm_core, and
unprotected_aterm is renamed to unprotected_aterm_core. This is done in
this commit.

The next step is to rename aterm_appl to aterm.

After that, all occurrences of aterm_core must be removed, generally
by replacing them by aterm, which is the former aterm_appl.
Ultimately, the current class aterm should be removed.

All tests compile and header tests succeed.