Replaced every aterm_appl by an aterm.
In the process of removing the class aterm and replacing
it with aterm_appl, we have renamed aterm_appl to aterm.
In the previous commit all occurrences of aterm were renamed
to aterm_core.

All tests succeed at the moment.

The next task is to remove all occurrences of aterm_core as
they should not be needed anymore. It is estimated that
at the moment there are approximately 1500 of them in the
toolset.