Resolve problem #1737, where lpsxsim would crash
Lpsxsim would crash when loading a trace smaller than the
trace in the existing window. This would invoke updateSimulation
where inside updateSimulation the size of the visual trace in the right
part of the window would be changed. In case the focus would fall
outside of this window, this would call to change the focus, which
in turn would call updateSimulation. The problem is now solved
by not allowing the recursively call updateSimulation using a
simple boolean.