Incorporated the 64 bit representation for Nat and Pos.
It is now possible to translate the toolset with 64 bit positive and
natural numbers, as well as using it with the positive and natural
numbers as specified in Groote and Mousavi, Modelling and Analysing
Communicating Systems, The MIT Press, 2014.

The choice is made by defining Enable64bitNumbers in the file
mcrl2/data/data_configuration.h.

When the flag is set, the files pos64.spec and nat64.spec will
be used to generate the definition for positive and natural numbers.
The results are found in pos64.h and nat64.h. All other generated
data types also use the 64 bit extension, because they depend on
the 64 bit numbers, with the exception of a function update.

Throughout the code, Enable64bitNumbers is used to select either
the old or the new data types where appropriate.

All tests succeed, both with and without the flag.

The pos.spec and nat.spec files with rewrite rules