Adapted the new rewriter to avoid declaring too many aterm variables.
The manually implemented functions for machine words and also
function updates used the stack to pass results. This has now been
adapted such that they write the result in addresses avoiding to
have intermediate addresses being put in a protection set very
frequently.

Furthermore, made a copy of the size of a rewrite stack, to
avoid calling the function size() of an atermpp::vector very
frequently, as this function goes through the busy-forbidden
protocol each time size() is called.